Allows us to not rely on Guard to run a test
Guard with the guard minitest helps us have minitest in our testing
environment.

The spec helper provides that dependency for anyone that simply wants to
run a test, without relying on a tool they may not be familiar with, or
not in need of a continuous testing environment for that moment.