Keep default file structure for 'rake' command
The 'rake' command expects the file structure to be a certain way in our
gem.  We could modify where 'rake' looks for the test files but then
we'd have to override where 'rake' looks for our tests.

I'm sure we could override the pathing 'rake' uses to get to our tests
but I couldn't figure it out so I'm using this file sctructure instead.